refactor(#154): 폴더 위치 수정, facade 통합
