+++ ai/src/main/java/dukku/ai/global/advisor/MemoryExtractionAdvisor.java
@@ -1,4 +1,4 @@
-package dukku.ai.global.config.advisor;
+package dukku.ai.global.advisor;
 
 import org.springframework.ai.chat.client.ChatClientRequest;
 import org.springframework.ai.chat.client.ChatClientResponse;
